Use upstream git completion but oh-my-zsh prompts
The gitfast plugin lets me use the zsh completion scripts that come from
upstream git, which are faster.  But it also forces me into using the
upstream prompt logic, which isn't as nice as oh-my-zsh's.  I added an
option to the gitfast plugin that lets me opt out of the upstream prompt
info.