Make SSH from tmux refresh gpg-agent tty
Instead of refreshing the gpg-agent tty once at the beginning of each
login, I now do this every time I call `ssh`.  This makes sure that if I
call `ssh` from within a tmux session, and the connection needs to grab
a key from the gpg-agent, the pinentry is shown inside of that tmux
session, and not in the tty of the inbound connection where I happened
to have called `tmux attach`.