Move refreshing ssh to separate file
I have an ssh wrapper that refreshes my gpg-agent's tty, so that any
pinentry dialog shows up on the same screen that I'm calling ssh from.
I originally implemented this as a zsh function, but I need it to be a
proper executable script, so that I can set things like $GIT_SSH to
force git to use the wrapper instead of calling /usr/bin/ssh directly.